update implementation and tests
clear calendar

fix failing tests

format

use internal stage array binding parameter for test

ignore on Github actions

move parameter query
