+++ src/test/java/net/snowflake/client/jdbc/BindUploaderIT.java
@@ -62,9 +62,9 @@ public class BindUploaderIT extends BaseJDBCTest {
   };
 
   static final String csv1 =
-      "42,1234,12.34,12.34,42,row1,807F,1970-01-01,00:00:00.000000000,1970-01-01 00:00:00.000000000 +00:00";
+      "42,1234,12.34,12.34,42,row1,807F,1970-01-01,00:00:00.000000000,1970-01-01 00:00:00.000000000 Z";
   static final String csv2 =
-      "420,12340,120.34,120.34,420,row2,7F80,1970-01-01,00:00:00.001000000,1970-01-01 00:00:00.001000000 +00:00";
+      "420,12340,120.34,120.34,420,row2,7F80,1970-01-01,00:00:00.001000000,1970-01-01 00:00:00.001000000 Z";
 
   static final String STAGE_DIR = "binduploaderit";
   static final String SELECT_FROM_STAGE =
@@ -203,67 +203,5 @@ public class BindUploaderIT extends BaseJDBCTest {
     list.add("another bind value");
     map.put("1", new ParameterBindingDTO("", list));
     assertEquals(2, BindUploader.arrayBindValueCount(map));
-  }
-
-  // Test setting the input stream buffer size to be quite small and ensure that 2 files (1
-  // corresponding to each input stream) are created in internal stage.
-  @Test
-  public void testUploadedResultsMultiple() throws Exception {
-
-    // Get an estimate of how many bytes are in 1 of the rows being uploaded and set this value to
-    // be the input stream buffer size. For 2 similarly sized rows, we should now have 2 files.
-    int lengthOfOneRow = csv1.getBytes("UTF-8").length;
-    bindUploader.setInputStreamBufferSize(lengthOfOneRow);
-    bindUploader.upload(getBindings(conn));
-    // assert 2 files were created on internal stage
-    assertEquals(2, bindUploader.getFileCount());
-    Statement stmt = conn.createStatement();
-    // assert that the results look proper
-    ResultSet rs = stmt.executeQuery(SELECT_FROM_STAGE);
-    rs.next();
-    assertEquals(csv1, parseRow(rs));
-    rs.next();
-    assertEquals(csv2, parseRow(rs));
-    assertFalse(rs.next());
-    rs.close();
-    stmt.close();
-  }
-
-  // Test single csv upload and successful parsing
-  @Test
-  public void testUploadedResultsSimple() throws Exception {
-    bindUploader.upload(getBindings(conn));
-
-    Statement stmt = conn.createStatement();
-    ResultSet rs = stmt.executeQuery(SELECT_FROM_STAGE);
-    rs.next();
-    assertEquals(csv1, parseRow(rs));
-    rs.next();
-    assertEquals(csv2, parseRow(rs));
-    assertFalse(rs.next());
-    rs.close();
-    stmt.close();
-  }
-
-  @Test
-  public void testUploadStreamLargeBatch() throws Exception {
-    // create large batch so total bytes transferred are about 10 times the size of input stream
-    // buffer
-    int batchSize = 1024 * 1024;
-    SnowflakePreparedStatementV1 stmt =
-        (SnowflakePreparedStatementV1) conn.prepareStatement(dummyInsert);
-    for (int i = 0; i < batchSize; i++) {
-      bind(stmt, row1);
-    }
-    Map<String, ParameterBindingDTO> parameterBindings = stmt.getBatchParameterBindings();
-    bindUploader.upload(parameterBindings);
-    ResultSet rs = stmt.executeQuery(SELECT_FROM_STAGE);
-    for (int i = 0; i < batchSize; i++) {
-      rs.next();
-      assertEquals(csv1, parseRow(rs));
-    }
-    assertFalse(rs.next());
-    rs.close();
-    stmt.close();
   }
 }
